Changed sequence ID numbers in the cb_link_to_coarse data structure to be represented with 64-bit integers and added more documentation.

+++ src/link_to_coarse.c
@@ -10,7 +10,7 @@
 /*Initializes a link to a sequence in the coarse database.  Creates the
   link's edit script using the alignment passed in.*/
 struct cb_link_to_coarse *
-cb_link_to_coarse_init(int32_t coarse_seq_id,
+cb_link_to_coarse_init(uint64_t coarse_seq_id,
                        uint64_t original_start, uint64_t original_end,
                        uint16_t coarse_start, uint16_t coarse_end,
                        struct cb_alignment alignment, bool dir){
@@ -35,7 +35,7 @@ cb_link_to_coarse_init(int32_t coarse_seq_id,
 /*Initializes a link to a coarse sequence in which the coarse and original
   sequences are identical and therefore a blank edit script can be used.*/
 struct cb_link_to_coarse *
-cb_link_to_coarse_init_nodiff(int32_t coarse_seq_id,
+cb_link_to_coarse_init_nodiff(uint64_t coarse_seq_id,
                               uint64_t original_start, uint64_t original_end,
                               uint16_t coarse_start, uint16_t coarse_end,
                               bool dir){
